+++ src/components/Header.tsx
@@ -11,6 +11,7 @@ import { getMemberInfo, userLogout } from '@src/api/authApi'
 import { removeCookieToken } from '@src/storage/Cookie'
 import { useEffect } from 'react'
 import { SET_INFO, DELETE_INFO } from '@src/store/slices/infoSlice'
+import PATH from '@src/constants/pathConst'
 
 type PropsType = {
   setSideOpen: React.Dispatch<React.SetStateAction<boolean>>
@@ -76,9 +77,9 @@ const Header = ({ setSideOpen }: PropsType) => {
             </Link>
             <div className="menu">
               <Link to="/help">고객센터</Link>
-              {authenticated && <Link to="/mypage">마이페이지</Link>}
-              {!authenticated && <Link to="/join">회원가입</Link>}
-              {authenticated && <Link to="/one_on_one">1:1 문의</Link>}
+              {authenticated && <Link to={PATH.MYPAGE}>마이페이지</Link>}
+              {!authenticated && <Link to={PATH.JOIN}>회원가입</Link>}
+              {authenticated && <Link to={PATH.ASK}>1:1 문의</Link>}
               {authenticated ? <a onClick={logoutHandler}>로그아웃</a> : <Link to="/login">로그인</Link>}
               <button
                 onClick={() => {
