+++ src/pages/Ask.tsx
@@ -9,6 +9,7 @@ import { getQuestion, getAdminQuestion } from '@src/api/getApi'
 import { useSelector } from 'react-redux'
 import { RootState } from '@src/store/config'
 import OneOneOne from '@components/OneOnOne'
+import PATH from '@src/constants/pathConst'
 
 const Ask = () => {
   const [questions, setQuestions] = useState<QuestionGetTypes[]>([])
@@ -54,7 +55,7 @@ const Ask = () => {
           </QuestionContainer>
           <OtherAskStyle>
             <span>원하는 답변이 없다면?</span>
-            <button onClick={() => navigate('/help_form')}>1:1 질문하기</button>
+            <button onClick={() => navigate(PATH.HELP_FORM)}>1:1 질문하기</button>
           </OtherAskStyle>
         </Container>
       ) : (
@@ -77,7 +78,7 @@ const Ask = () => {
           </QuestionContainer>
           <MobileOtherAsk>
             <span>원하는 답변이 없다면?</span>
-            <button onClick={() => navigate('/help_form')}>1:1 질문하기</button>
+            <button onClick={() => navigate(PATH.HELP_FORM)}>1:1 질문하기</button>
           </MobileOtherAsk>
         </Inner>
       )}
